+++ packages/api/src/modules/push-notifications/push-notifications.controller.ts
@@ -1,19 +1,16 @@
 import { Body, Controller, Get, Post } from '@nestjs/common';
 import { PushSubscription } from 'web-push';
 
-import { TokenPayload } from '../../interfaces/token-types';
-import { AuthService } from '../auth/auth.service';
 import { Auth } from '../auth/decorators/auth.decorator';
 import { CookiePayload } from '../auth/decorators/cookiePayload.decorator';
+import { UserFromCookiePayloadPipe } from '../auth/pipes/user-from-cookie-payload.pipe';
+import { UserEntity } from '../database/entities/user.entity';
 import { PushNotificationsService } from './push-notifications.service';
 
 @Auth()
 @Controller('push-notifications')
 export class PushNotificationsController {
-  constructor(
-    private readonly pushNotificationsService: PushNotificationsService,
-    private readonly authService: AuthService,
-  ) {}
+  constructor(private readonly pushNotificationsService: PushNotificationsService) {}
 
   @Get()
   async send(): Promise<void> {
@@ -25,10 +22,9 @@ export class PushNotificationsController {
 
   @Post()
   async subscribe(
-    @CookiePayload() { sub }: TokenPayload,
+    @CookiePayload(UserFromCookiePayloadPipe) user: UserEntity,
     @Body() subscription: PushSubscription,
   ): Promise<void> {
-    const user = await this.authService.getUser(sub);
     await this.pushNotificationsService.subscribe({ subscription, user });
   }
 }