+++ packages/client/src/elements/index.ts
@@ -20,4 +20,7 @@ export { default as SelectCard } from './inputs/SelectCard';
 export { CardLayout, DefaultLayout, GlobalLayout, TabbedLayout } from './layouts';
 export { default as Link } from './Link';
 export { CardList, DetailedList } from './lists';
+export { default as Portal } from './Portal';
 export { default as Snackbar } from './Snackbar';
+export { default as ToggleSlider } from './ToggleSlider';
+export { default as Tooltip } from './Tooltip';